fix: 컨트롤러, 서비스 역할/책임 분리
