+++ src/main/java/com/sadowbass/outerpark/application/reservation/dto/PendingId.java
@@ -1,0 +1,11 @@
+package com.sadowbass.outerpark.application.reservation.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class PendingId {
+
+    private String pendingId;
+}
